backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==3.1.1
Django==5.1.0
django-cors-headers==4.9.0
django-recaptcha==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
soupsieve==2.7
sqlparse==0.5.5
typing_extensions==4.13.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.11.0
